requirements.txt
openai-agents>=0.2.4
tavily-python>=0.7.10
python-dotenv>=1.0.0 